jwt_proxy/requirements.txt
flask
pyjwt
flask-cors
ijson
waitress

